+++ R/crew_launcher.R
@@ -6,6 +6,7 @@
 #'   `@inheritParams crew::crew_launcher` in the source code file of
 #'   [crew_launcher_local()].
 #' @inheritParams crew_client
+#' @inheritParams crew_worker
 #' @param name Name of the launcher.
 #' @param seconds_interval Number of seconds between
 #'   polling intervals waiting for certain internal
@@ -69,6 +70,10 @@
 #'   Plugins that may use these processes should run asynchronous calls
 #'   using `launcher$async$eval()` and expect a `mirai` task object
 #'   as the return value.
+#' @param r_arguments Optional character vector of command line arguments
+#'   to pass to `Rscript` (non-Windows) or `Rscript.exe` (Windows)
+#'   when starting a worker. Example:
+#'   `r_arguments = c("--vanilla", "--max-connections=32")`.
 #' @examples
 #' if (identical(Sys.getenv("CREW_EXAMPLES"), "true")) {
 #' client <- crew_client()
@@ -97,7 +102,9 @@ crew_launcher <- function(
   garbage_collection = FALSE,
   launch_max = 5L,
   tls = crew::crew_tls(),
-  processes = NULL
+  processes = NULL,
+  r_arguments = c("--no-save", "--no-restore"),
+  options_metrics = crew::crew_options_metrics()
 ) {
   crew_deprecate(
     name = "seconds_exit",
@@ -127,7 +134,9 @@ crew_launcher <- function(
     garbage_collection = garbage_collection,
     launch_max = launch_max,
     tls = tls,
-    processes = processes
+    processes = processes,
+    r_arguments = r_arguments,
+    options_metrics = options_metrics
   )
 }
 
@@ -169,6 +178,8 @@ crew_class_launcher <- R6::R6Class(
     .launch_max = NULL,
     .tls = NULL,
     .processes = NULL,
+    .r_arguments = NULL,
+    .options_metrics = NULL,
     .async = NULL,
     .throttle = NULL
   ),
@@ -238,6 +249,14 @@ crew_class_launcher <- R6::R6Class(
     processes = function() {
       .subset2(private, ".processes")
     },
+    #' @field r_arguments See [crew_launcher()].
+    r_arguments = function() {
+      .subset2(private, ".r_arguments")
+    },
+    #' @field options_metrics See [crew_launcher()].
+    options_metrics = function() {
+      .subset2(private, ".options_metrics")
+    },
     #' @field async A [crew_async()] object to run low-level launcher tasks
     #'   asynchronously.
     async = function() {
@@ -267,6 +286,8 @@ crew_class_launcher <- R6::R6Class(
     #' @param launch_max See [crew_launcher()].
     #' @param tls See [crew_launcher()].
     #' @param processes See [crew_launcher()].
+    #' @param r_arguments See [crew_launcher()].
+    #' @param options_metrics See [crew_launcher()].
     #' @examples
     #' if (identical(Sys.getenv("CREW_EXAMPLES"), "true")) {
     #' client <- crew_client()
@@ -295,7 +316,9 @@ crew_class_launcher <- R6::R6Class(
       garbage_collection = NULL,
       launch_max = NULL,
       tls = NULL,
-      processes = NULL
+      processes = NULL,
+      r_arguments = NULL,
+      options_metrics = NULL
     ) {
       private$.name <- name
       private$.seconds_interval <- seconds_interval
@@ -312,6 +335,8 @@ crew_class_launcher <- R6::R6Class(
       private$.launch_max <- launch_max
       private$.tls <- tls
       private$.processes <- processes
+      private$.r_arguments <- r_arguments
+      private$.options_metrics <- options_metrics
     },
     #' @description Validate the launcher.
     #' @return `NULL` (invisibly).
@@ -405,6 +430,12 @@ crew_class_launcher <- R6::R6Class(
         crew_assert(identical(colnames(private$.workers), cols))
         crew_assert(nrow(private$.workers) > 0L)
       }
+      if (!is.null(private$.r_arguments)) {
+        crew_assert(
+          is.character(private$.r_arguments),
+          message = "r_arguments must be a character vector."
+        )
+      }
       crew_assert(
         inherits(private$.tls, "crew_class_tls"),
         message = "field 'tls' must be an object created by crew_tls()"
@@ -423,6 +454,9 @@ crew_class_launcher <- R6::R6Class(
           message = "field 'throttle' must be an object from crew_throttle()"
         )
         private$.throttle$validate()
+      }
+      if (!is.null(private$.options_metrics)) {
+        crew_options_metrics_validate(private$.options_metrics)
       }
       invisible()
     },
@@ -444,6 +478,7 @@ crew_class_launcher <- R6::R6Class(
         (8L * as.integer(isTRUE(private$.garbage_collection)))
       list(
         url = socket,
+        asyncdial = FALSE,
         autoexit = crew_terminate_signal(),
         cleanup = cleanup,
         output = TRUE,
@@ -479,13 +514,20 @@ crew_class_launcher <- R6::R6Class(
           settings = settings,
           launcher = launcher,
           worker = worker,
-          instance = instance
+          instance = instance,
+          options_metrics = crew::crew_options_metrics(
+            path = path,
+            seconds_interval = seconds_interval
+          )
         ),
         env = list(
           settings = self$settings(socket),
           launcher = launcher,
           worker = worker,
-          instance = instance
+          instance = instance,
+          path = private$.options_metrics$path,
+          seconds_interval = private$.options_metrics$seconds_interval %|||%
+            5
         )
       )
       out <- deparse_safe(expr = call, collapse = " ")
@@ -696,7 +738,9 @@ crew_class_launcher <- R6::R6Class(
       crew_assert(
         futile <= private$.launch_max,
         message = paste0(
-          "{crew} worker ",
+          "{crew} worker with name ",
+          name,
+          " and index ",
           index,
           " launched ",
           private$.launch_max,