+++ inst/doc/logging.R
@@ -1,7 +1,7 @@
 ## ----include = FALSE----------------------------------------------------------
+library(crew)
 knitr::opts_chunk$set(
   collapse = TRUE,
   comment = "#>"
 )
-library(crew)